agent/nodes/requirement.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.43.1
anyio==4.8.0
asttokens==3.0.0
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
Chroma==0.2.0
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
copilotkit==0.1.33
cryptography==43.0.3
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.1.0
faiss-cpu==1.9.0.post1
fastapi==0.115.6
fastembed==0.5.1
filelock==3.16.1
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
groq==0.15.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.25.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.14
langchain-anthropic==0.3.3
langchain-community==0.3.14
langchain-core==0.3.30
langchain-groq==0.2.3
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langgraph==0.2.64
langgraph-api==0.0.16
langgraph-checkpoint==2.0.10
langgraph-cli==0.1.67
langgraph-sdk==0.1.51
langsmith==0.2.11
loguru==0.7.3
marshmallow==3.25.1
matplotlib-inline==0.1.7
mmh3==4.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.1
onnxruntime==1.20.1
openai==1.59.8
orjson==3.10.15
packaging==24.2
parso==0.8.4
partialjson==0.0.8
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.3
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.41.3
structlog==24.4.0
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
yarl==1.18.3
